Fredrick-Updated workflow for nextflow analysis
