+++ Original_code/input/A1_01_01_1_reference_Exon6.fasta
@@ -1,0 +1,2 @@
+>ABO*A1.01.01.1
+TAGGAAGGATGTCCTCGTGGTGACCCCTTGGCTGGCTCCCATTGTCTGGGAGGGCACATTCAACATCGACATCCTCAACGAGCAGTTCAGGCTCCAGAACACCACCATTGGGTTAACTGTGTTTGCCATCAAGAA
