+++ README.md
@@ -6,3 +6,8 @@ ABO sequences were aquired from the NCBI dbRBC database:
 https://www.ncbi.nlm.nih.gov/projects/gv/mhc/xslcgi.cgi?cmd=bgmut/home
 
 
+# The nextflow workflow
+
+run `nextflow run main.nf -resume`
+
+
