+++ assets/multiqc_config.yaml
@@ -1,0 +1,41 @@
+report_comment: >
+  This report has been generated by the <a href="https://github.com/fmobegi/readaccuracy" target="_blank">fmobegi/readaccuracy</a>
+  analysis pipeline.
+  For software versions used in this analysis, please check the `software_versions.txt` file in the output directory
+  
+report_section_order:
+  software_versions:
+    order: -1000
+  "software_versions.txt":
+    order: -1001
+
+export_plots: true
+
+# Run only these modules
+run_modules:
+  - software_versions
+  - fastqc
+  - samtools
+ 
+# Order of modules
+module_order:
+  - fastqc:
+      name: "FastQC (raw)"
+      path_filters:
+        - "*.zip"
+  - samtools:
+      name: "Samtools Flagstat"
+      
+extra_fn_clean_exts:
+  - "_val"
+
+# Don't show % Dups in the General Stats table (we have this from Picard)
+table_columns_visible:
+  fastqc:
+    percent_duplicates: False
+
+sp:
+  samtools/stats:
+    fn: "*.samtools.stats"
+  samtools/flagstat:
+    fn: "*.samtools.flagstat"
